@@ -1,7 +1,7 @@
 package com.google.ar.sceneform.rendering;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import com.google.android.filament.Entity;
 import com.google.android.filament.IndexBuffer;
 import com.google.android.filament.VertexBuffer;
