@@ -1,6 +1,6 @@
 package com.google.ar.sceneform.rendering;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import com.google.android.filament.Material;
 
 public class MaterialInternalDataGltfImpl extends MaterialInternalData {
