@@ -1,6 +1,6 @@
 package com.google.ar.sceneform.utilities;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 /**
  * Static convenience methods that help a method or constructor check whether it was invoked
