@@ -15,8 +15,8 @@
  */
 package com.google.ar.sceneform.ux;
 
-import android.support.annotation.CallSuper;
-import android.support.annotation.Nullable;
+import androidx.annotation.CallSuper;
+import androidx.annotation.Nullable;
 import com.google.ar.sceneform.FrameTime;
 import com.google.ar.sceneform.Node;
 
