@@ -1,8 +1,8 @@
 package com.google.ar.sceneform;
 
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
 import android.view.MotionEvent;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
 import com.google.ar.core.Pose;
 
 import com.google.ar.sceneform.collision.Ray;
