@@ -1,7 +1,7 @@
 package com.google.ar.sceneform.rendering;
 
-import android.support.annotation.Nullable;
 import android.util.Log;
+import androidx.annotation.Nullable;
 import com.google.android.filament.Entity;
 import com.google.android.filament.EntityManager;
 import com.google.android.filament.LightManager;
