@@ -5,10 +5,10 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Picture;
 import android.graphics.PorterDuff;
-import android.support.annotation.Nullable;
 import android.view.Surface;
 import android.view.View;
 import android.widget.LinearLayout;
+import androidx.annotation.Nullable;
 
 import com.google.ar.sceneform.utilities.Preconditions;
 import java.util.ArrayList;
