@@ -1,8 +1,8 @@
 package com.google.ar.sceneform.rendering;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.util.Log;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import com.google.android.filament.Box;
 import com.google.android.filament.Entity;
 import com.google.android.filament.EntityInstance;
