@@ -2,11 +2,11 @@ package com.google.ar.sceneform.rendering;
 
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.Nullable;
-import android.support.annotation.RequiresApi;
-import android.support.annotation.VisibleForTesting;
 import android.view.Surface;
 import android.view.SurfaceView;
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
+import androidx.annotation.VisibleForTesting;
 import com.google.android.filament.Camera;
 import com.google.android.filament.Entity;
 import com.google.android.filament.IndirectLight;
