@@ -1,8 +1,8 @@
 package com.google.ar.sceneform.utilities;
 
 import android.content.res.AssetManager;
-import android.support.annotation.Nullable;
 import android.util.Log;
+import androidx.annotation.Nullable;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
