@@ -15,9 +15,9 @@
  */
 package com.google.ar.sceneform.ux;
 
-import android.support.annotation.Nullable;
 import android.util.DisplayMetrics;
 import android.view.MotionEvent;
+import androidx.annotation.Nullable;
 import com.google.ar.sceneform.HitTestResult;
 import java.util.ArrayList;
 
